small logging improvements
