+++ config.go
@@ -1,0 +1,6 @@
+package roc
+
+import "github.com/hashicorp/go-hclog"
+
+var LogLevel = hclog.Info
+
